+++ mojo_opset/backends/ttx/operators/position_embedding.py
@@ -1,3 +1,4 @@
+from typing import List
 from typing import Optional
 from typing import Tuple
 
@@ -6,8 +7,10 @@ import torch
 import mojo_opset.core.operators.position_embedding as mojo_position_embedding
 from mojo_opset.backends.ttx.kernels import rot_pos_embed
 from mojo_opset.backends.ttx.kernels import rope_fwd
+from mojo_opset.backends.ttx.kernels.npu.mrope import mrope_fwd_impl
 from mojo_opset.core import MojoRotaryEmbedding
 from mojo_opset.core import MojoApplyRoPE
+from mojo_opset.core.operators.position_embedding import MojoMRoPE
 
 
 class TTXRotaryEmbedding(MojoRotaryEmbedding):
@@ -48,6 +51,22 @@ class TTXApplyRoPE(MojoApplyRoPE):
         head_first: bool = True,
     ) -> Tuple[torch.Tensor, torch.Tensor]:
         return rope_fwd(q, k, cos, sin, head_first)
+
+
+class TTXMRoPE(MojoMRoPE):
+    supported_platforms_list = ["npu"]
+
+    @staticmethod
+    def forward(
+        q: torch.Tensor,
+        k: torch.Tensor,
+        cos: torch.Tensor,
+        sin: torch.Tensor,
+        mrope_section: List[int],
+        is_interleaved: bool = False,
+        head_dim: Optional[int] = None,
+    ) -> Tuple[torch.Tensor, torch.Tensor]:
+        return mrope_fwd_impl(q, k, cos, sin, mrope_section, is_interleaved)
 
 
 def _patch_torch_apply_rope_for_batched_cos() -> None:
